App - adjust module
